+++ UXperience/Controller/DetailView.swift
@@ -114,10 +114,3 @@ extension DetailView: UITableViewDelegate, UITableViewDataSource {
     }
 
 }
-
-protocol DetailViewDelegate: AnyObject {
-    func goToWebSite()-> String
-    func titleOfNews()-> String
-    func setFrontCardText()-> String
-    func setFrontCardTitle()-> String
-}
