+++ UXperience/Protocol/CardCollectionViewDelegate.swift
@@ -1,0 +1,12 @@
+//
+//  CardCollectionViewDelegate.swift
+//  UXperience
+//
+//  Created by Luiz Sena on 05/12/22.
+//
+
+import Foundation
+
+protocol CardCollectionViewDelegate: AnyObject {
+    func teste(with viewModel: LawsModel)
+}
