+++ UXperience/Protocol/DetailViewDelegate.swift
@@ -1,0 +1,15 @@
+//
+//  DetailViewDelegate.swift
+//  UXperience
+//
+//  Created by Luiz Sena on 05/12/22.
+//
+
+import Foundation
+
+protocol DetailViewDelegate: AnyObject {
+    func goToWebSite()-> String
+    func titleOfNews()-> String
+    func setFrontCardTitle()-> String
+    func setFrontCardText()-> String
+}
